Create 02_create_wp_db.sh
This is part 2 of the script , it will prompt for the db root password you used during the secure install, name of the database that you want to create, the WordPress admin username that you want to give and the password that you want to use for that user. At the end of this it will read back what inputs you gave it.